--- apps/back/src/interceptors/response.interceptor.ts
@@ -1,24 +1,0 @@
-import {
-  CallHandler,
-  ExecutionContext,
-  Injectable,
-  NestInterceptor,
-} from '@nestjs/common';
-import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
-
-@Injectable()
-export class ResponseInterceptor implements NestInterceptor {
-  intercept(context: ExecutionContext, next: CallHandler): Observable<any> {
-    return next.handle().pipe(
-      map((data) => {
-        const response = context.switchToHttp().getResponse();
-
-        if ('statusCode' in data && typeof data.statusCode === 'number') {
-          response.status(data.statusCode);
-        }
-        return data;
-      }),
-    );
-  }
-}
